refactor: split local installer responsibilities
